docs: Fix locale switcher in mixed routing example when logged in
