+++ examples/example-app-router-mixed-routing/src/i18n/request.ts
@@ -1,18 +1,20 @@
 import {hasLocale} from 'next-intl';
 import {getRequestConfig} from 'next-intl/server';
-import {defaultLocale, locales} from '../config';
-import {getUserLocale} from '../db';
+import {COOKIE_NAME, defaultLocale, locales} from '../config';
+import {cookies} from 'next/headers';
 
 export default getRequestConfig(async ({requestLocale}) => {
   // Read from potential `[locale]` segment
+  // if the user is on a public page
   let candidate = await requestLocale;
 
   if (!candidate) {
-    // The user is logged in
-    candidate = await getUserLocale();
+    // Read from cookie if the user is logged in
+    const store = await cookies();
+    candidate = store.get(COOKIE_NAME)?.value;
   }
-  const locale = hasLocale(locales, candidate) ? candidate : defaultLocale;
 
+  const locale = hasLocale(locales, candidate) ? candidate : defaultLocale;
   return {
     locale,
     messages: (await import(`../../messages/${locale}.json`)).default
